Move identity functions from useAuthStatus to identity file
